refactor: now default page is about. Soon it will be the store
